figuring out simulation
